Moving workflow logic to a controller
