+++ src/app.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import workflowsRouter from './routers/workflowsRouter'
+import workflowsRouter from './routers/workflowRouter'
 const app = express();
 //TODO: Move midlewares setup to separate files under middlewares folder
 express.json();
